Continue rewriting to JNI.
Now functions and imports are supported.

I also rewrite some of the C code and did more transformations in the
Java world. Its all done in the NativeAdapter, which converts a lot of
values before passing them to JNI and wise versa.

All unit tests are go green now! *happy*