Introduce docker based build and testing infrastructure.
To build the native libs, we now use a docker container.
This make it easer for others to rebuild the native libs.
The build container run on CentOS 5 (glibc 2.5) to increase the
compatibility of the native libs.

The checks are now run on multiple docker containers, running
- CentOS 5
- CentOS 6
- CentOS 7
- ubuntu 12.04
- ubuntu 14.04

Now we can guarantee that jsass will run on a wide range of linux based
operating systems, running glibc 2.5 or newer.

To run the build processes we provide gradle tasks.
Use `gradle buildNativeLibs` to build the native libs.
Use `gradle check` to run the unittests on the testing containers
(including checkstyle, pmd and local junit tests).